+++ lib/ruuuby/class/nums/int.rb
@@ -9,7 +9,7 @@
 # [ℚ]  the mathematical rational-numbers(+ℚ+) (all ints can be represented as a rational)
 # [ℂ]  the mathematical complex-numbers(+ℂ+) (all ints can be represented as complex numbers {}with imaginary component of 0})
 # [ℝ]  the mathematical real-numbers(+ℝ+)
-# [𝕌]  the mathematical universal-set(+ℝ+) (in this context only excludes (NaN, +∞, and -∞) which ints may not be and/or represent)
+# [𝕌]  the mathematical universal-set(+𝕌+) (in this context only excludes (NaN, +∞, and -∞) which ints may not be and/or represent)
 class ::Integer
 
   # ---------------------------------------------------------------------------------------------------------- | *f11* |
@@ -17,12 +17,12 @@ class ::Integer
   # @return [Boolean] true, all *non-negative* integers(+ℤ+) are considered in the set of natural-numbers(+ℕ+)
   def ℕ? ; self >= 0; end
 
-  # ∀ x ∈ Rℤ, x ∈ ℤ
+  # ∀ x ∈ Rℤ --> x ∈ ℤ
   #
   # ∴ ⨍(`infinite?`) will always return `false`
   ⨍_add_aliases(:infinite?, [:∞?, :♾️?, :∞ℂ?, :nan?, :¿?])
 
-  # ∀ x ∈ Rℤ, x ∈ ℤ, x ∈ ℚ, x ∈ ℂ, x ∈ ℝ, x ∈ 𝕌
+  # ∀ x ∈ Rℤ --> x ∈ ℤ, x ∈ ℚ, x ∈ ℂ, x ∈ ℝ, x ∈ 𝕌
   #
   # ∴ ⨍(`finite?`) will always return `true`
   ⨍_add_aliases(:finite?, [:ℤ?, :ℚ?, :ℂ?, :ℝ?, :𝕌?])