+++ spec/db/db_spec.rb
@@ -18,4 +18,30 @@ RSpec.describe 'db/db.rb' do
 
   end
 
+  #                   __          ___     __
+  #   /\     |  |    |  \    |     |     /__`
+  #  /~~\    \__/    |__/    |     |     .__/
+  context 'audits', :'audits' do
+
+    context 'has correct db configs' do
+      context 'has correct directory structure' do
+        context 'directory{db}' do
+          it 'exists' do
+            expect(🗄️.∃?('./db')).to eq(true)
+          end
+          context 'directory{db/seeds}' do
+            it 'exists' do
+              expect(🗄️.∃?('./db/seeds')).to eq(true)
+            end
+          end
+        end
+        it 'directory{db/migrate} exists' do
+          expect(🗄️.∃?('./db/migrate')).to eq(true)
+          expect(📁.∃?('./db/migrate.rb')).to eq(false)
+        end
+      end
+    end
+
+  end
+
 end